datahub-chatbot-backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic